Fix typo(?) in match_zurl()
The return value of 2 doesn't make much sense to me
as the function is used in a conditional like
if(!match_zurl(...)), thus should return True and False
to make a noticeable difference.

In a practical setup, the old code meant that pazpar2 would
not correctly work when using multiple SRU and Z39.50 servers
with a settings file using a wildcard for each of those.

The resulting error messages looked like pazpar2 was trying
to send Z39.50-style queries to the SRU servers

19:19:43-13/10 pazpar2 [warn] Failed to generate CQL query, code=-1
19:19:43-13/10 pazpar2 [log] PQF: @attr 1=1016 bread

and possibly vice-versa, depending on the order of the
settings in configuration files.