add logic for multipart works and series titles
Background:
Volumes can be multipart works inside a series.
This can be explicitly modelled in Pica data using the 036C/D and 036E/F fields.
As Marc only has the 490(/830) fields for this purpose without indicators for the hierarchy, it is difficult to tell these apart.
If there is a multipart work inside a series, the guideline is to put the multipart work title in the _first_ 490 field.
Thus: figure out whether this is a multipart work using position 19 in the leader (values b and c).
If it is a multipart work use the first of the 490 fields as the new multivolume-title and subsequent ones as the series title.

Example:
http://gso.gbv.de/sru/DB=2.1?query=pica.ppn+%3D+%22594671094%22&version=1.1&operation=searchRetrieve&recordSchema=marc21&maximumRecords=10&startRecord=1&recordPacking=xml